Split utils into lists and maps
